update: use fetch instead alchemy sdk

+++ index.js
@@ -1,10 +1,12 @@
-import fs from "fs";
+import fs from "fs-extra";
 import {
+  getCollectionInfo,
   getNftsForCollection,
   getOwnerForCollection,
   combineTo721,
 } from "./lib.js";
-import { intro, outro, text } from "@clack/prompts";
+import { intro, outro, text, confirm, cancel } from "@clack/prompts";
+import { exit } from "node:process";
 
 async function main() {
   intro("NFT Collection snapshot");
@@ -18,18 +20,35 @@ async function main() {
     },
   });
 
+  const contractInfo = await getCollectionInfo(contractAddress);
+  const checked = await confirm({
+    message: `Is the collection name ${contractInfo.name}?`,
+  });
+
+  if (!checked) {
+    cancel("Canceled.");
+    exit(0);
+  }
+
   try {
     const owners = await getOwnerForCollection(contractAddress);
     const result = await getNftsForCollection(contractAddress);
 
+    if (result == null) {
+      outro("No NFTs found.");
+      return;
+    }
+
+    fs.removeSync("result");
+    fs.mkdirSync("result");
     fs.writeFileSync(
-      "./result.json",
+      "result/collection.json",
       // TODO: now supports only ERC721
-      JSON.stringify(combineTo721(owners, result)),
+      JSON.stringify(await combineTo721(owners, result)),
       "utf8"
     );
 
-    outro("done!\nSee: result.json");
+    outro("Done!🎉 See result directory.");
   } catch (e) {
     console.error(e);
   }
